LayoutTests/fast/forms/suggested-value-after-setvalue-expected.txt

This test setting a value clears suggested value.

After setting suggestedValue:
| <input>
|   id="test"
|   type="text"
|   value="initial value"
|   this.value="initial value"
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "suggested value"
| "input.value: initial value"
| "internals.suggestedValue(input): suggested value"
| "input.selectionStart: 0"
| "input.selectionEnd: 0"

After setting value:
| <input>
|   id="test"
|   type="text"
|   value="initial value"
|   this.value="new value"
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "new value"
| "input.value: new value"
| "internals.suggestedValue(input): "
| "input.selectionStart: 9"
| "input.selectionEnd: 9"
| "PASS"
